Only use override options on case/contact managment handler when explicity set
Support the enable override configuration option, so the override settings will
only be used when explicity set.
Adds a new update hook which will set this setting to 'on' for all webforms
where they had already set an override for the post_url and auth_header
(this is so parking forms automatically update, until they are added as a form
group).